Introduce QueryQueueStack
