+++ java/cdk/src/main/java/sleeper/cdk/stack/WebSocketQueryStack.java
@@ -36,8 +36,6 @@ import software.amazon.awscdk.services.lambda.IFunction;
 import software.amazon.awscdk.services.lambda.Permission;
 import software.amazon.awscdk.services.s3.Bucket;
 import software.amazon.awscdk.services.s3.IBucket;
-import software.amazon.awscdk.services.sqs.IQueue;
-import software.amazon.awscdk.services.sqs.Queue;
 import software.constructs.Construct;
 
 import sleeper.cdk.Utils;
@@ -52,7 +50,6 @@ import java.util.Locale;
 import java.util.Map;
 
 import static sleeper.cdk.Utils.createLambdaLogGroup;
-import static sleeper.configuration.properties.instance.CdkDefinedInstanceProperty.QUERY_QUEUE_ARN;
 import static sleeper.configuration.properties.instance.CommonProperty.ID;
 
 public class WebSocketQueryStack extends NestedStack {
@@ -64,14 +61,13 @@ public class WebSocketQueryStack extends NestedStack {
     public WebSocketQueryStack(Construct scope,
             String id,
             InstanceProperties instanceProperties,
-            BuiltJars jars, CoreStacks coreStacks) {
+            BuiltJars jars, CoreStacks coreStacks, QueryQueueStack queryQueueStack) {
         super(scope, id);
 
         IBucket jarsBucket = Bucket.fromBucketName(this, "JarsBucket", jars.bucketName());
         LambdaCode queryJar = jars.lambdaCode(BuiltJar.QUERY, jarsBucket);
         setupWebSocketApi(instanceProperties, queryJar, coreStacks);
-        IQueue queryQueue = Queue.fromQueueArn(this, id, instanceProperties.get(QUERY_QUEUE_ARN));
-        queryQueue.grantSendMessages(webSocketApiHandler);
+        queryQueueStack.grantSendMessages(webSocketApiHandler);
         Utils.addStackTagIfSet(this, instanceProperties);
     }
